Use relative imports throughout tests
The tests relied on Python2 relative import semantics. Python3 changed
import semantics to always search sys.path by default. To import using a
relative path it must have a leading dot.

Forward compatible with newer Pythons.

Works towards the goal of moving tests outside of the installed package.

For more information, see PEP-328:

https://www.python.org/dev/peps/pep-0328/

+++ tests/test_types_basic.py
@@ -26,9 +26,9 @@ import decimal
 
 import sys
 from functools import wraps
-import testutils
+from . import testutils
 import unittest
-from testutils import ConnectingTestCase, decorate_all_tests
+from .testutils import ConnectingTestCase, decorate_all_tests
 
 import psycopg2
 
@@ -326,7 +326,7 @@ class TypesBasicTests(ConnectingTestCase):
         self.assertEqual(1, f1)
         i1 = self.execute("select -%s;", (-1,))
         self.assertEqual(1, i1)
-        l1 = self.execute("select -%s;", (-1L,))
+        l1 = self.execute("select -%s;", (long(-1),))
         self.assertEqual(1, l1)
 
     def testGenericArray(self):
